Consolidate PSPropertyExpressionFilter to eliminate code duplication (#25043)
