+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/FormatAndOutput/format-object/Format-Object.cs
@@ -76,7 +76,7 @@ namespace Microsoft.PowerShell.Commands
 
             if (ExcludeProperty != null)
             {
-                parameters.excludePropertyFilter = new Microsoft.PowerShell.Commands.Internal.Format.PSPropertyExpressionFilter(ExcludeProperty);
+                parameters.excludePropertyFilter = new PSPropertyExpressionFilter(ExcludeProperty);
 
                 // ExcludeProperty implies -Property * for better UX
                 if (_props == null || _props.Length == 0)
